requirements.txt
blinker==1.6.2
click==8.1.3
docutils==0.20.1
Flask==2.3.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
lockfile==0.12.2
MarkupSafe==2.1.3
numpy==1.25.0
opencv-python==4.8.0.74
python-daemon==3.0.1
PyYAML==6.0
Werkzeug==2.3.6
zipp==3.15.0
